# docs/api/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "N11 Credit Application Center API",
    "version": "1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/users/{nationalIdNumber}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getByNationalIdNumber",
        "parameters": [
          {
            "name": "nationalIdNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "user-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "nationalIdNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user-controller"
        ],
        "operationId": "deleteByNationalIdNumber",
        "parameters": [
          {
            "name": "nationalIdNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credit-applications": {
      "post": {
        "tags": [
          "credit-application-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditApplicationRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getAllUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credit-applications/": {
      "get": {
        "tags": [
          "credit-application-controller"
        ],
        "operationId": "getByByNationalIdNumberAndBirthDate",
        "parameters": [
          {
            "name": "birthdate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserRequestDto": {
        "type": "object",
        "properties": {
          "nationalIdNumber": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        }
      },
      "CreditApplicationRequestDto": {
        "type": "object",
        "properties": {
          "nationalIdNumber": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "assurance": {
            "type": "number",
            "format": "double"
          },
          "monthlyIncome": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}